Fix host ssif driver unable to create /dev/ipmi0 interface
When host boots up, during the ssif enumeration, it will send request
commands to BMC to query for data then bring up the /dev/ipmi0 interface.
The timeout duration to wait for data of host is 15 seconds, if timeouted,
host will send other commands and continue the enumeration.

During the AC or DC cycle, BMC dbus will be busy with a lot of traffic
which is easy to cause the dbus timeout happened. Ipmid comunicates with
ssifbridge via dbus. When ipmid receive requests, it puts in a queue and
response to those requests one by one. Because of the dbus timeout, ipmid
will send the response late which cause the mismatch and duplicate
responses in the request of host. Let say, host request command 87,
but because of the timeout, ipmid response with command 66 which may be
sent by host previously. And then, ipmid sends response of command 87.
This leads to 2 or more reponses will be sent to host.

These mismatches and duplicate responses will cause host driver confused
and failed to create ssif interface on host.

Tested:
1. Run AC cycle 200 times, ensure that ipmi in-band commands still work
2. Run DC cycle 400 times, ensure that ipmi in-band commands still work

Signed-off-by: quang.ampere <quangn@amperecomputing.com>
Change-Id: Ie08a6dd376a1460b3e63c7ecb3908ac886c521b0